Moved React components to a new `components` folder
