+++ client_src/javascripts/index.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './App';
+import App from './components/App';
 
 /*
  * Retrieve the data from the head when the page has been loaded.
